moved WorkflowViewModelBase and WorkflowServiceExtension to samples
