+++ notes.txt
@@ -1,3 +1,10 @@
+notes:
+- move to samples/WebApi
+  (x) WorkflowViewModelBase.cs 
+  (x) WorkflowServiceExtension.cs 
+
+
+----------------------------------------------------------------------------------------------------
 EF commands:
 - see: https://docs.microsoft.com/en-us/ef/core/miscellaneous/cli/dotnet
