+++ samples/WebApi/Workflows/Holiday/HolidayService.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using tomware.Microwf.Core;
 using tomware.Microwf.Engine;
 using WebApi.Domain;
+using WebApi.Common;
 
 namespace WebApi.Workflows.Holiday
 {
