+++ samples/WebApi/Workflows/Issue/IssueController.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using tomware.Microwf.Engine;
+using WebApi.Common;
 
 namespace WebApi.Workflows.Issue
 {
